separate test file for invalid sym operand
